requirements.txt
amazon-textract-caller
amazon-textract-prettyprinter
amazon-textract-response-parser
amazon-textract-textractor
boto3==1.36.23
botocore==1.36.23
python-dotenv==1.0.1
